Bugfix: ImportError following some codestyle/comment updates
